"refactor: ТЗ11 - исправлена реализация сервисов и хранилищ"

+++ src/main/java/ru/yandex/practicum/filmorate/controllers/FilmController.java
@@ -13,6 +13,7 @@ import ru.yandex.practicum.filmorate.service.film.FilmService;
 
 import javax.validation.Valid;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @RestController
 @RequestMapping("/films")
@@ -28,8 +29,8 @@ public class FilmController {
     }
 
     @GetMapping
-    public List<Film> filmList() {
-        return filmStorage.findAllFilms();
+    public List<Film> filmList() throws ValidationException {
+        return filmService.findAllFilms();
     }
 
     @GetMapping("/{id}")
@@ -37,20 +38,20 @@ public class FilmController {
         if (id <= 0) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
-        return filmStorage.getFilmById(id);
+        return filmService.getFilmById(id);
     }
 
     @ResponseBody
     @PostMapping
     public Film create(@Valid @RequestBody Film film) throws ValidationException {
-        film = filmStorage.createFilm(film);
+        film = filmService.createFilm(film);
         return film;
     }
 
     @ResponseBody
     @PutMapping
     public Film update(@Valid @RequestBody Film film) throws ValidationException {
-        film = filmStorage.updateFilm(film);
+        film = filmService.updateFilm(film);
         return film;
     }
 
@@ -61,16 +62,19 @@ public class FilmController {
 
     @DeleteMapping("/{id}/like/{userId}")
     public void deleteLike(@PathVariable int id, @PathVariable int userId) throws ValidationException {
+        if (id < 1 || userId < 1) {
+            throw new NoSuchElementException();
+        }
         filmService.deleteLike(id, userId);
     }
 
     @DeleteMapping("/{id}")
-    public Film delete(@PathVariable int id) throws ValidationException {
-        return filmStorage.delete(id);
+    public void delete(@PathVariable int id) throws ValidationException {
+        filmStorage.delete(id);
     }
 
     @GetMapping("/popular")
-    public List<Film> getPopular(@RequestParam(name = "count", defaultValue = "10") Integer count) {
+    public List<Film> getPopular(@RequestParam(name = "count", defaultValue = "10") Integer count) throws ValidationException {
         return filmService.getPopular(count);
     }
 }
